# backend/openapi.yaml
openapi: 3.1.0
info:
  title: WhatsApp AI Agent — Multi-Tenant SaaS API
  version: 1.0.0
  description: |
    ## Overview
    Multi-tenant restaurant ordering platform powered by an AI agent over WhatsApp.
    Each **Business** (tenant) connects a WhatsApp phone number, and the AI agent
    handles customer orders, menu inquiries, complaints, and status checks automatically.

    **Two API surfaces:**
    | Surface | Auth | Scope |
    |---------|------|-------|
    | **Tenant API** | `x-api-key` or Bearer JWT | Scoped to one business |
    | **Platform API** | Bearer JWT only (`businessId: null`) | Cross-tenant super-admin |

    ## Webhook Flow
    1. Meta sends a WhatsApp message to `POST /api/webhook`
    2. Server identifies the Business via `phone_number_id`
    3. AI agent processes the message (menu query, order submission, etc.)
    4. Reply is sent back as the response body

    ## Real-time Events (Socket.io)
    - **URL:** `http://localhost:3001`
    - **Rooms:** `business:{businessId}`
    - **Events:**
      | Event | Payload | Description |
      |-------|---------|-------------|
      | `new-message` | `{ customerId, customerPhone, message, reply }` | AI replied to a customer |
      | `new-order` | `{ orderId, referenceId, customerId, items, totalPrice }` | New order placed via WhatsApp |
      | `order-updated` | `{ orderId, status, updatedAt }` | Order status changed |
...

    ## Demo Credentials
    | Role | API Key / Endpoint | Email | Password |
...
servers:
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Business API key (e.g. `demo-api-key-123`)
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: User access token from `/api/auth/login`
    PlatformBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Platform user JWT from `/api/auth/platform-login`.
        Token contains `businessId: null`.
  schemas:
...
    MenuCategory:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        nameAr:
          type: string
          nullable: true
        sortOrder:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/MenuItem'
    MenuItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        nameAr:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        price:
          type: number
        available:
          type: boolean
        categoryId:
          type: string
    MenuCategoryInput:
      type: object
      properties:
        name:
          type: string
        nameAr:
          type: string
        sortOrder:
          type: integer
    MenuItemInput:
      type: object
      required: [name, price, categoryId]
      properties:
        name:
          type: string
        nameAr:
          type: string
        description:
          type: string
        price:
          type: number
        categoryId:
          type: string
        available:
          type: boolean

    # ─── Orders ───
    Order:
      type: object
      properties:
        id:
          type: string
        referenceId:
          type: string
          example: ORD-000001
        businessId:
          type: string
        customer:
          $ref: '#/components/schemas/Customer'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        totalPrice:
          type: number
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, delivered, cancelled]
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OrderItem:
      type: object
      properties:
        id:
          type: string
        menuItem:
          $ref: '#/components/schemas/MenuItem'
        quantity:
          type: integer
        notes:
          type: string
          nullable: true
    StatusUpdateInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, delivered, cancelled]

    # ─── Customers & Conversations ───
    Customer:
      type: object
      properties:
        id:
          type: string
        phone:
          type: string
        name:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /menu:
    get:
      tags: [Menu]
      summary: Get full menu with categories and items
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '200':
          description: Menu categories with nested items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MenuCategory'
  /menu/search:
    get:
      tags: [Menu]
      summary: Search menu items
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Matching menu items (max 20)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MenuItem'
        '400':
          description: Missing search query
  /menu/categories:
    post:
      tags: [Menu]
      summary: Create a menu category
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MenuCategoryInput'
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuCategory'
  /menu/categories/{id}:
    put:
      tags: [Menu]
      summary: Update a menu category
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MenuCategoryInput'
      responses:
        '200':
          description: Category updated
        '404':
          description: Category not found
  /menu/items:
    post:
      tags: [Menu]
      summary: Create a menu item
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MenuItemInput'
      responses:
        '201':
          description: Item created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuItem'
        '400':
          description: Invalid price or category
  /menu/items/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Menu]
      summary: Get menu item by ID
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '200':
          description: Menu item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuItem'
        '404':
          description: Not found
    put:
      tags: [Menu]
      summary: Update a menu item
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MenuItemInput'
      responses:
        '200':
          description: Item updated
        '404':
          description: Not found
    delete:
      tags: [Menu]
      summary: Delete a menu item
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '200':
          description: Item deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '404':
          description: Not found
  /menu/items/{id}/toggle:
    patch:
      tags: [Menu]
      summary: Toggle menu item availability
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Item availability toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuItem'
        '404':
          description: Not found

  # ─── ORDERS ──────────────────────────────────────────────────────────────────
  /orders:
    get:
      tags: [Orders]
      summary: List orders with optional status filter
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, preparing, ready, delivered, cancelled]
      responses:
        '200':
          description: Orders list with customer and items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
  /orders/{id}:
    get:
      tags: [Orders]
      summary: Get order details
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order with items and customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '403':
          description: Order belongs to a different business
        '404':
          description: Order not found
  /orders/{id}/status:
    put:
      tags: [Orders]
      summary: Update order status
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusUpdateInput'
      responses:
        '200':
          description: Status updated, Socket.io event emitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid status
        '404':
          description: Order not found

  # ─── SETTINGS ────────────────────────────────────────────────────────────────
...